Add security checks to controllers, update authentication type

+++ teamfinder/src/backend/java/hu/elte/teamfinder/controllers/AccountController.java
@@ -13,8 +13,9 @@ import hu.elte.teamfinder.viewmodels.AccountViewModel;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+import org.springframework.security.core.Authentication;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.util.MimeTypeUtils;
 import org.springframework.web.bind.annotation.*;
@@ -22,7 +23,6 @@ import org.springframework.web.bind.annotation.*;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
-import java.security.Principal;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
@@ -45,8 +45,6 @@ public class AccountController {
         this.profileService = profileService;
     }
 
-    // TODO: add @PreAuthorize for permission based authentication for each Mapping
-
     /*
     /**
      * Getting an account by ID
@@ -64,13 +62,12 @@ public class AccountController {
     */
 
     @GetMapping("/all")
-    public ResponseEntity<?> getAllAccounts(Principal principal) {
-        // UsernamePasswordAuthenticationToken authenticationToken = principal.;
-        /*
-        if(!accountDetails.getAuthorities().contains(new SimpleGrantedAuthority("ADMIN")))
-        {
+    public ResponseEntity<?> getAllAccounts(Authentication authentication) {
+
+        if (!isAdmin(authentication)) {
             return new ResponseEntity<String>("Access denied", FORBIDDEN);
-        }*/
+        }
+
         List<AccountViewModel> accountViewModels = new ArrayList<>();
 
         List<Account> accounts = accountService.getAllAccounts();
@@ -80,9 +77,13 @@ public class AccountController {
     }
 
     @GetMapping("/get/{id}")
-    public ResponseEntity<AccountViewModel> getAccountById(@PathVariable("id") Long id) {
+    public ResponseEntity<?> getAccountById(
+            @PathVariable("id") Long id, Authentication authentication) {
+        if (!isMeOrAdmin(id, authentication)) {
+            return new ResponseEntity<String>("Access denied", FORBIDDEN);
+        }
         Account account = accountService.getAccountById(id);
-        return new ResponseEntity<>(new AccountViewModel(account), HttpStatus.OK);
+        return new ResponseEntity<AccountViewModel>(new AccountViewModel(account), HttpStatus.OK);
     }
 
     @PostMapping("/add")
@@ -102,8 +103,13 @@ public class AccountController {
 
     @PutMapping("/update/{id}")
     public ResponseEntity<?> updateAccount(
-            @PathVariable("id") Long id, @RequestBody Account account) {
-        // TODO: Implement function
+            @PathVariable("id") Long id,
+            @RequestBody Account account,
+            Authentication authentication) {
+
+        if (!isMeOrAdmin(id, authentication)) {
+            return new ResponseEntity<String>("Access denied", FORBIDDEN);
+        }
         Account foundAccount;
         try {
             foundAccount = accountService.getAccountById(id);
@@ -123,7 +129,11 @@ public class AccountController {
     }
 
     @DeleteMapping("/delete/{id}")
-    public ResponseEntity<?> deleteAccountById(@PathVariable("id") Long id) {
+    public ResponseEntity<?> deleteAccountById(
+            @PathVariable("id") Long id, Authentication authentication) {
+        if (!isMeOrAdmin(id, authentication)) {
+            return new ResponseEntity<String>("Access denied", FORBIDDEN);
+        }
         accountService.deleteAccount(id);
         return new ResponseEntity<>(HttpStatus.OK);
     }
@@ -132,7 +142,7 @@ public class AccountController {
     public void refreshToken(HttpServletRequest request, HttpServletResponse response)
             throws IOException {
         // TODO: maybe make member
-        JwtUtil jwtUtil = new JwtUtil();
+        JwtUtil jwtUtil = new JwtUtil(accountService);
 
         String authorizationHeader = request.getHeader(AUTHORIZATION);
         if (authorizationHeader != null && authorizationHeader.startsWith(JwtUtil.HEADER_PREFIX)) {
@@ -161,5 +171,26 @@ public class AccountController {
                 new ObjectMapper().writeValue(response.getOutputStream(), error);
             }
         }
+    }
+
+    private boolean isMeOrAdmin(Long id, Authentication authentication) {
+        AccountDetails accountDetails = (AccountDetails) authentication.getPrincipal();
+        // Only allow if user is updating their own account data or if user is admin
+        if (accountDetails.getAccountId() == id
+                || accountDetails
+                        .getAuthorities()
+                        .contains((new SimpleGrantedAuthority("ROLE_ADMIN")))) {
+            return true;
+        }
+        return false;
+    }
+
+    private boolean isAdmin(Authentication authentication) {
+        // Only allow fetching all accounts if user is admin
+        AccountDetails accountDetails = (AccountDetails) authentication.getPrincipal();
+        if (accountDetails.getAuthorities().contains((new SimpleGrantedAuthority("ROLE_ADMIN")))) {
+            return true;
+        }
+        return false;
     }
 }
